+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.20.2 - 2025-08-07
+
 - fix: missing session_id property for snapshot events ([#273](https://github.com/PostHog/posthog-android/pull/273))
 
 ## 3.20.1 - 2025-07-28